+++ History.md
@@ -1,3 +1,5 @@
+* BMurri: WIXBUG:4032 - Enable heat to accomodate COM relative paths and foreign servers.
+
 * BobArnson: WIXBUG:3807 - Add AssemblyFoldersEx key.
 
 * BobArnson: WIXBUG:3737 - Block with specific message attempt to use RemotePayload outside ExePackage.